backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlmodel>=0.0.24
sqlalchemy>=2.0
pydantic>=2.10
pydantic-settings>=2.6
python-multipart>=0.0.12
python-dotenv>=1.0
httpx>=0.28
pillow>=10.4
huggingface_hub>=0.30
PyJWT>=2.10
email-validator>=2.1
